Move html functions to `string::html`
