+++ lib/fuzz/fuzz_targets/fuzz_executor.dict
@@ -218,10 +218,6 @@
 "hash"
 "geo::hash::decode("
 "geo::hash::encode("
-"html"
-"html::"
-"html::encode("
-"html::sanitize("
 "http"
 "http::"
 "http::head("
@@ -298,6 +294,8 @@
 "string::distance::hamming("
 "string::distance::levenshtein("
 "string::endsWith("
+"string::html::encode("
+"string::html::sanitize("
 "string::is::alphanum("
 "string::is::alpha("
 "string::is::ascii("
