+++ lib/fuzz/fuzz_targets/fuzz_sql_parser.dict
@@ -217,10 +217,6 @@
 "hash"
 "geo::hash::decode("
 "geo::hash::encode("
-"html"
-"html::"
-"html::encode("
-"html::sanitize("
 "http"
 "http::"
 "http::head("
@@ -296,6 +292,8 @@
 "string::distance::hamming("
 "string::distance::levenshtein("
 "string::endsWith("
+"string::html::encode("
+"string::html::sanitize("
 "string::is::alphanum("
 "string::is::alpha("
 "string::is::ascii("
